[#136457191] Don't secure token; drop unused ssl setting
The app didn't work in production because the `token` was coming through
as `undefined` so the API (to us) failed with `forbidden`

This change makes the `token` `secure: false`, which means the client
app can use it. As far as we can tell there is not an issue with this
change given our use case.